move helper methods out of core.rs
